[refactor]userServiceのフォルダ移動
