+++ server/src/application/user/user.controller.spec.ts
@@ -1,6 +1,6 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { UserController } from './user.controller';
-import { UserService } from '../../domain/user/user.service';
+import { UserService } from '../../domain/user/service/user.service';
 import { getRepositoryToken } from '@nestjs/typeorm';
 import { User } from '../../domain/entities/user.entity';
 import { Following } from '../../domain/entities/following.entity';
